+++ proto/agent_grpc.pb.go
@@ -19,20 +19,21 @@ import (
 const _ = grpc.SupportPackageIsVersion9
 
 const (
-	AwgAgent_Init_FullMethodName           = "/awgmesh.AwgAgent/Init"
-	AwgAgent_RotateToken_FullMethodName    = "/awgmesh.AwgAgent/RotateToken"
-	AwgAgent_AddTunnel_FullMethodName      = "/awgmesh.AwgAgent/AddTunnel"
-	AwgAgent_RemoveTunnel_FullMethodName   = "/awgmesh.AwgAgent/RemoveTunnel"
-	AwgAgent_ListTunnels_FullMethodName    = "/awgmesh.AwgAgent/ListTunnels"
-	AwgAgent_AddPeer_FullMethodName        = "/awgmesh.AwgAgent/AddPeer"
-	AwgAgent_RemovePeer_FullMethodName     = "/awgmesh.AwgAgent/RemovePeer"
-	AwgAgent_ListPeers_FullMethodName      = "/awgmesh.AwgAgent/ListPeers"
-	AwgAgent_RotateParams_FullMethodName   = "/awgmesh.AwgAgent/RotateParams"
-	AwgAgent_GetParams_FullMethodName      = "/awgmesh.AwgAgent/GetParams"
-	AwgAgent_CaptureRefresh_FullMethodName = "/awgmesh.AwgAgent/CaptureRefresh"
-	AwgAgent_GetStatus_FullMethodName      = "/awgmesh.AwgAgent/GetStatus"
-	AwgAgent_GetRoutes_FullMethodName      = "/awgmesh.AwgAgent/GetRoutes"
-	AwgAgent_GetHealth_FullMethodName      = "/awgmesh.AwgAgent/GetHealth"
+	AwgAgent_Init_FullMethodName             = "/awgmesh.AwgAgent/Init"
+	AwgAgent_RotateToken_FullMethodName      = "/awgmesh.AwgAgent/RotateToken"
+	AwgAgent_AddTunnel_FullMethodName        = "/awgmesh.AwgAgent/AddTunnel"
+	AwgAgent_RemoveTunnel_FullMethodName     = "/awgmesh.AwgAgent/RemoveTunnel"
+	AwgAgent_ListTunnels_FullMethodName      = "/awgmesh.AwgAgent/ListTunnels"
+	AwgAgent_UpdateTunnelPeer_FullMethodName = "/awgmesh.AwgAgent/UpdateTunnelPeer"
+	AwgAgent_AddPeer_FullMethodName          = "/awgmesh.AwgAgent/AddPeer"
+	AwgAgent_RemovePeer_FullMethodName       = "/awgmesh.AwgAgent/RemovePeer"
+	AwgAgent_ListPeers_FullMethodName        = "/awgmesh.AwgAgent/ListPeers"
+	AwgAgent_RotateParams_FullMethodName     = "/awgmesh.AwgAgent/RotateParams"
+	AwgAgent_GetParams_FullMethodName        = "/awgmesh.AwgAgent/GetParams"
+	AwgAgent_CaptureRefresh_FullMethodName   = "/awgmesh.AwgAgent/CaptureRefresh"
+	AwgAgent_GetStatus_FullMethodName        = "/awgmesh.AwgAgent/GetStatus"
+	AwgAgent_GetRoutes_FullMethodName        = "/awgmesh.AwgAgent/GetRoutes"
+	AwgAgent_GetHealth_FullMethodName        = "/awgmesh.AwgAgent/GetHealth"
 )
 
 // AwgAgentClient is the client API for AwgAgent service.
@@ -46,6 +47,7 @@ type AwgAgentClient interface {
 	AddTunnel(ctx context.Context, in *AddTunnelRequest, opts ...grpc.CallOption) (*AddTunnelResponse, error)
 	RemoveTunnel(ctx context.Context, in *RemoveTunnelRequest, opts ...grpc.CallOption) (*RemoveTunnelResponse, error)
 	ListTunnels(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*TunnelList, error)
+	UpdateTunnelPeer(ctx context.Context, in *UpdateTunnelPeerRequest, opts ...grpc.CallOption) (*UpdateTunnelPeerResponse, error)
 	// Peer management (endpoint mode)
 	AddPeer(ctx context.Context, in *AddPeerRequest, opts ...grpc.CallOption) (*AddPeerResponse, error)
 	RemovePeer(ctx context.Context, in *RemovePeerRequest, opts ...grpc.CallOption) (*RemovePeerResponse, error)
@@ -113,6 +115,16 @@ func (c *awgAgentClient) ListTunnels(ctx context.Context, in *Empty, opts ...grp
 	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(TunnelList)
 	err := c.cc.Invoke(ctx, AwgAgent_ListTunnels_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *awgAgentClient) UpdateTunnelPeer(ctx context.Context, in *UpdateTunnelPeerRequest, opts ...grpc.CallOption) (*UpdateTunnelPeerResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(UpdateTunnelPeerResponse)
+	err := c.cc.Invoke(ctx, AwgAgent_UpdateTunnelPeer_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -220,6 +232,7 @@ type AwgAgentServer interface {
 	AddTunnel(context.Context, *AddTunnelRequest) (*AddTunnelResponse, error)
 	RemoveTunnel(context.Context, *RemoveTunnelRequest) (*RemoveTunnelResponse, error)
 	ListTunnels(context.Context, *Empty) (*TunnelList, error)
+	UpdateTunnelPeer(context.Context, *UpdateTunnelPeerRequest) (*UpdateTunnelPeerResponse, error)
 	// Peer management (endpoint mode)
 	AddPeer(context.Context, *AddPeerRequest) (*AddPeerResponse, error)
 	RemovePeer(context.Context, *RemovePeerRequest) (*RemovePeerResponse, error)
@@ -257,6 +270,9 @@ func (UnimplementedAwgAgentServer) RemoveTunnel(context.Context, *RemoveTunnelRe
 }
 func (UnimplementedAwgAgentServer) ListTunnels(context.Context, *Empty) (*TunnelList, error) {
 	return nil, status.Error(codes.Unimplemented, "method ListTunnels not implemented")
+}
+func (UnimplementedAwgAgentServer) UpdateTunnelPeer(context.Context, *UpdateTunnelPeerRequest) (*UpdateTunnelPeerResponse, error) {
+	return nil, status.Error(codes.Unimplemented, "method UpdateTunnelPeer not implemented")
 }
 func (UnimplementedAwgAgentServer) AddPeer(context.Context, *AddPeerRequest) (*AddPeerResponse, error) {
 	return nil, status.Error(codes.Unimplemented, "method AddPeer not implemented")
@@ -392,6 +408,24 @@ func _AwgAgent_ListTunnels_Handler(srv interface{}, ctx context.Context, dec fun
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(AwgAgentServer).ListTunnels(ctx, req.(*Empty))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _AwgAgent_UpdateTunnelPeer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(UpdateTunnelPeerRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(AwgAgentServer).UpdateTunnelPeer(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: AwgAgent_UpdateTunnelPeer_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(AwgAgentServer).UpdateTunnelPeer(ctx, req.(*UpdateTunnelPeerRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -584,6 +618,10 @@ var AwgAgent_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "ListTunnels",
 			Handler:    _AwgAgent_ListTunnels_Handler,
+		},
+		{
+			MethodName: "UpdateTunnelPeer",
+			Handler:    _AwgAgent_UpdateTunnelPeer_Handler,
 		},
 		{
 			MethodName: "AddPeer",